requirements.txt
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.7
django-allauth==0.39.1
django-ckeditor==5.8.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-heroku==0.3.1
django-import-export==1.2.0
django-jet==1.0.8
django-js-asset==1.2.2
django-storages==1.8
django-tinymce==2.8.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
google-api-core==1.14.3
google-auth==1.8.2
google-cloud-core==1.1.0
google-cloud-storage==1.23.0
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
gunicorn==20.0.0
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==3.0.1
pep8==1.7.1
Pillow==6.2.1
protobuf==3.11.1
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
python-decouple==3.1
python-dotenv==0.10.3
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.13.0
sqlparse==0.3.0
stripe==2.27.0
tablib==0.14.0
urllib3==1.25.7
whitenoise==4.1.4
xlrd==1.2.0
xlwt==1.3.0
